Updated obstacles and added ceiling
New obstacles have been added that are smaller but move up and down (bouncing off top and bottom) at varying speeds. They also spawn at various heights. A ceiling was added that kills the player. This is because the new obstacle must have a longer scoring trigger and the obstacle has a stopping height. This meant the player could fly above the camera, not be able to be hit by the obstacle, and still hit the score trigger.